django/requirements.txt
asgiref==3.4.0
autopep8==1.5.7
Django==3.2.4
djangorestframework==3.12.4
drf-yasg==1.20.0
dj-database-url==0.5.0
gunicorn==20.1.0
psycopg2==2.9.1
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
